move tee  commands
